Changed URL of Santa Cruz Resistance Community.
The old 2-step method (with a public info page on G+ and then a "PR"
account) was confusing and led to a lot of people not getting any
response from us. The new link is directly to the G+ page of the Santa
Cruz Resistance account. The page will eventually be done away with.